http_worker is a library component
